Extract gdbus helpers into their own file
